Puts all the base converters together
Puts all the base converters together in a single code with conditional statements (from decimal to binary, hexadecimal, and octal number systems).

The base value for conversion is to be input by the user and assigned to a variable. The equivalent conversion code will run based on the value of the "base" variable.